refactor(ws): split monolithic mod.rs into focused files
